+++ modules/mod_nxpeasyforms/services/provider.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * @package     NXP Easy Forms
+ * @subpackage  mod_nxpeasyforms
+ * @copyright   Copyright (C) 2024-2025 nexusplugins.com. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+declare(strict_types=1);
+
+\defined('_JEXEC') or die;
+
+use Joomla\CMS\Extension\Service\Provider\HelperFactory;
+use Joomla\CMS\Extension\Service\Provider\Module;
+use Joomla\CMS\Extension\Service\Provider\ModuleDispatcherFactory;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+
+return new class () implements ServiceProviderInterface {
+    /**
+     * Registers the service provider with a DI container.
+     */
+    public function register(Container $container): void
+    {
+        $container->registerServiceProvider(new ModuleDispatcherFactory('\\Joomla\\Module\\Nxpeasyforms'));
+        $container->registerServiceProvider(new HelperFactory('\\Joomla\\Module\\Nxpeasyforms\\Site\\Helper'));
+
+        $container->registerServiceProvider(new Module());
+    }
+};